feat: Implement LSP server with vscode-wasm
