+++ tsconfig.json
@@ -8,6 +8,6 @@
     "sourceMap": true,
     "esModuleInterop": true
   },
-  "include": ["src/*"],
+  "include": ["src/**/*"],
   "exclude": ["node_modules"]
 }
